requirements.txt
jinja2
ops
lightkube
lightkube-models
pytest-interface-tester
pydantic
